Add mkdir command to DLS Makefile
The version of setuptools being used at DLS on RHEL 7 no longer creates
the install directory by itself (as the RHEL 6 version did).
This must be done before easy_install is called.

Use "dls-python --version" to determine the short python
version so that the specific python version is not hard coded.